third_party/WebKit/LayoutTests/fast/html/imports/import-success-fail.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<head>
<link id="toSuccess" rel="import" href="resources/hello.html">
<link id="toFail" rel="import" href="resources/no-such-file.html">
<link id="notImport" rel="stylesheet" href="resources/no-such-file.css">
</head>
<body>
<script>
shouldBe("toSuccess.import.querySelector('h1').innerHTML", "'Hello'");
shouldBeNonNull("toSuccess.import.querySelector('html')");
shouldBeNull("toFail.import");
shouldBeNull("notImport.import");
</script>
</body>
</html>
