third_party/WebKit/LayoutTests/fast/html/imports/import-success-fail-expected.txt

PASS toSuccess.import.querySelector('h1').innerHTML is 'Hello'
PASS toSuccess.import.querySelector('html') is non-null.
PASS toFail.import is null
PASS notImport.import is null
PASS successfullyParsed is true

TEST COMPLETE
